docs: fix strange incompatibility with treesitter highlighting
When opening the readme in neovim 0.10.0, I got this error:

```
   Error  Error in decoration provider treesitter/highlighter.win:
Error executing lua: /usr/local/share/nvim/runtime/lua/vim/treesitter/query.lua:252: no such language: TOML
stack traceback:
```

Not sure why this is happening, but I'm guessing it's because of a bug
in the treesitter highlighting. It seems to be possible to work around
this by lowercasing the language name in the code block.