fix: endless loop when locations file is not found
If the .bacon-locations file is missing, the `Bacon.bacon_load()`
function goes into an endless loop, forever trying to recurse up path
the root directory (`/` -> `/../` -> `/../../` -> ...).

If I add a `vim.notify` call to the `Bacon.bacon_load()` function,
it shows notifications for `path`, `path/../` and `path/../../` but
never finishes.

```
   Info  11:33:29 notify.info found file: ".bacon-locations"
   Info  11:33:29 notify.info found file: "../"
   Info  11:33:29 notify.info found file: "../.bacon-locations"
   Info  11:33:29 notify.info found file: "../../"
```

This change seems to get rid of the endless loop. I tested this on nvim
0.10.0 and 0.9.0